+++ src/Window_SDL.c
@@ -258,7 +258,7 @@ static void OnWindowEvent(const SDL_Event* e) {
 		}
 }
 
-void Window_ProcessEvents(double delta) {
+void Window_ProcessEvents(float delta) {
 	SDL_Event e;
 	while (SDL_PollEvent(&e)) {
 		switch (e.type) {
@@ -295,7 +295,7 @@ void Window_ProcessEvents(double delta) {
 	}
 }
 
-void Window_ProcessGamepads(double delta) { }
+void Window_ProcessGamepads(float delta) { }
 
 static void Cursor_GetRawPos(int* x, int* y) {
 	SDL_GetMouseState(x, y);
